# src/interfaces/http/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node API contacts app",
    "version": "v1"
  },
  "servers": [
    {
      "description": "Local server",
      "url": "/api"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "operationId": "listContacts",
        "tags": [ "Contacts" ],
        "responses": {
          "200": {
            "description": "List of all contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createContact",
        "tags": [ "Contacts" ],
        "requestBody": {
          "description": "Contact data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewContact"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Contact not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "operationId": "showContact",
        "tags": [ "Contacts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of contact to show",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Return contact with given id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateContact",
        "tags": [ "Contacts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of contact to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "requestBody": {
          "description": "Contact new data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateContact"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteContact",
        "tags": [ "Contacts" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of contact to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "202": {
            "description": "Contact deleted successfully"
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contact": {
        "allOf": [
          { "$ref": "#/components/schemas/NewContact" },
          {
            "required": [ "id" ],
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ]
      },
      "NewContact": {
        "required": [ "name","email","phoneNumbers" ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumbers": {
            "type": "array"
          }
        }
      },
      "updateContact": {
        "required": [ "name","email" ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "ValidationError" ]
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        }
      },
      "ValidationErrorDetail": {
        "properties": {
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "NotFoundError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "NotFoundError" ]
          },
          "details": {
            "type": "string",
            "enum": [ "Contact with id {id} not found" ]
          }
        }
      }
    }
  }
}